requirements.txt
boto3==1.4.4
botocore==1.5.90
celery==4.1.0
Django==1.11.3
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-rest-swagger==2.1.2
djangorestframework==3.6.3
docutils==0.13.1
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
pytz==2017.2
requests==2.10.0
textacy==0.3.4
lxml==3.7.2
